requirements-gpu.txt
paddlepaddle-gpu==2.6.2.post120    # CUDA 12.0 (kiểm tra AWS instance)
paddleocr==2.7.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
PyMuPDF==1.23.8
pdf2image==1.16.3
numpy==1.24.4
pandas==2.1.4
openpyxl==3.1.2
vietocr==0.3.12
transformers==4.36.2
sentencepiece==0.1.99
torch==2.1.0+cu120                 # match CUDA version của AWS
tqdm==4.66.1
PyYAML==6.0.1
datasets==2.16.1
huggingface-hub==0.20.3